requirements-phase1.txt
# Phase 1: Core Infrastructure Dependencies
# Updated: October 29, 2025

# ===== Google ADK & Core Agent Frameworks =====
google-adk>=0.5.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# ===== LLM Providers =====
google-generativeai>=0.8.0
openai>=1.50.0
anthropic>=0.39.0
groq>=0.4.0

# ===== Agent Orchestration =====
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ===== State Management & Storage =====
redis>=5.0.0
sqlite-utils>=3.37

# ===== Configuration & Environment =====
python-dotenv>=1.0.0
pyyaml>=6.0.1
hydra-core>=1.3.2

# ===== Logging & Monitoring =====
structlog>=24.4.0
loguru>=0.7.2
colorama>=0.4.6

# ===== Async & Concurrency =====
asyncio>=3.4.3
aiohttp>=3.10.0
nest-asyncio>=1.6.0

# ===== Testing =====
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# ===== Development Tools =====
black>=24.8.0
ruff>=0.7.0
mypy>=1.11.0

# ===== Utilities =====
rich>=13.8.0
click>=8.1.7
tqdm>=4.66.0

# Note: Install with: pip install -r requirements-phase1.txt
# For development: pip install -r requirements-phase1.txt -e .

langgraph
mistralai
langchain-groq

neo4j
qdrant-client
sentence-transformers
pdf2image
Pillow
